requirements.txt
absl-py==1.4.0
appdirs==1.4.4
Bottleneck==1.3.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.0.4
cloudpickle==2.0.0
colorama==0.4.6
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.33
cytoolz==0.12.0
dask==2022.7.0
easydict==1.7
flit_core==3.6.0
fonttools==4.25.0
fsspec==2022.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
idna==3.4
imagecodecs==2021.8.26
imageio==2.19.3
joblib==1.1.1
json-tricks==3.16.1
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
networkx==2.8.4
ninja==1.11.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==22.0
pandas==1.5.2
partd==1.2.0
Pillow==9.3.0
pip==22.3.1
ply==3.11
pooch==1.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
setuptools==65.5.0
shapely==2.0.1
sip==6.6.2
six==1.16.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yacs==0.1.8
